[BOOKINGSG-9334][GZ] extract out components
